+++ cvxpy/atoms/affine/index.py
@@ -28,7 +28,7 @@ class index(AffAtom):
     # key - the index/slicing key (i.e. expr[key[0],key[1]]).
     def __init__(self, expr, key):
         # Format and validate key.
-        self.key = ku.validate_key(key, expr.shape)
+        self.key = ku.validate_key(key, expr._dcp_attr.shape)
         super(index, self).__init__(expr)
 
     # The string representation of the atom.
@@ -43,7 +43,7 @@ class index(AffAtom):
     def shape_from_args(self):
         """Returns the shape of the index expression.
         """
-        return u.Shape(*ku.size(self.key, self.args[0].shape))
+        return u.Shape(*ku.size(self.key, self.args[0]._dcp_attr.shape))
 
     def get_data(self):
         """Returns the (row slice, column slice).