+++ cvxpy/tests/test_problem.py
@@ -1246,3 +1246,11 @@ class TestProblem(BaseTest):
             self.assertTrue(np.allclose(x_alg, x_true, 1e-3))
             self.assertTrue(np.allclose(prob.value, np.linalg.norm(x_alg, p)))
             self.assertTrue(np.allclose(np.linalg.norm(x_alg, p), pnorm(x_alg, p).value))
+
+    def test_power(self):
+        x = Variable()
+        prob = Problem(Minimize(power(x, 1.7) + power(x, -2.3) - power(x, .45)))
+        prob.solve()
+        x = x.value
+        self.assertTrue(__builtins__['abs'](1.7*x**.7 - 2.3*x**-3.3 - .45*x**-.55) <= 1e-3)
+