+++ CHANGELOG.md
@@ -3,10 +3,8 @@ Version 1.0 (targets)
 * TODO add FAQ with info about solver errors, DCP errors, adding an objective in a for loop, dot product with numpy ND arrays.
 * TODO change *args to lists.
 * TODO change OPTIMAL to SOLVED?
-* TODO change SDP to only set upper triag == lower triag.
 * TODO separate out parts of DCPAttr and refactor so universal rules used for affine atoms.
 * TODO change size to size[0]*size[1] and shape to size.
-* TODO add integer/boolean variables.
 * TODO return 2D arrays instead of matrices.
 * TODO add sets and domains for Parameters/Variables that affect DCP properties.
 * TODO PSD parameters?
@@ -16,12 +14,14 @@ Version 1.0 (targets)
 * TODO make upper_tri_to_full atom?
 * TODO parameterize upper_tri (so takes elems above kth diagonal).
 * TODO replace norm with vecnorm and norm (for matrices) (see how numpy does it).
-* TODO redo tree_mat so it uses SMs (and doesn't compress the left hand side of A*x into a matrix).
 * TODO separate constant and linear components of expressions so potentially can cache matrix factorizations.
 * TODO add logistic_loss
 
 Version 0.2.21 (next release)
 -----------------------------
+* Made operator overloading work with scipy sparse matrices (with scipy 0.15).
+* Removed Expression shape function.
+* Removed Expression __array__ function.
 
 Version 0.2.20
 --------------